go mod tidy; move a test
