+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/google/go-github/v50 v50.2.0
 	github.com/pkg/errors v0.9.1
 	golang.org/x/mod v0.9.0
+	golang.org/x/oauth2 v0.6.0
 	golang.org/x/tools v0.7.0
 )
 
@@ -29,7 +30,6 @@ require (
 	github.com/xanzy/ssh-agent v0.3.3 // indirect
 	golang.org/x/crypto v0.7.0 // indirect
 	golang.org/x/net v0.8.0 // indirect
-	golang.org/x/oauth2 v0.6.0 // indirect
 	golang.org/x/sys v0.6.0 // indirect
 	google.golang.org/appengine v1.6.7 // indirect
 	google.golang.org/protobuf v1.28.0 // indirect
